# springboot-httpbin/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: dyrnq@outlook.com
    url: https://github.com/dyrnq
  description: springboot-httpbin
  title: springboot-httpbin
  version: 0.0.1
servers:
- url: https://httpbin.org/
...
paths:
...
  /basic-auth/{user}/{passwd}:
    get:
      parameters:
      - explode: false
        in: path
        name: user
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: passwd
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content: {}
          description: Sucessful authentication.
        "401":
          content: {}
          description: Unsuccessful authentication.
      summary: Prompts the user for authorization using HTTP Basic Auth.
      tags:
      - Auth
      x-accepts:
      - application/json
      x-tags:
      - tag: Auth
  /bearer:
    get:
      parameters:
      - explode: false
        in: header
        name: Authorization
        required: false
        schema:
          type: string
        style: simple
      - explode: false
        in: query
        name: token
        required: false
        schema:
          default: 3bsvhis2g6re31r0occao8i7ph
          type: string
        style: simple
      responses:
        "200":
          content: {}
          description: Sucessful authentication.
        "401":
          content: {}
          description: Unsuccessful authentication.
      summary: Prompts the user for authorization using bearer authentication.
      tags:
      - Auth
      x-accepts:
      - application/json
      x-tags:
      - tag: Auth
...
  /digest-auth/{qop}/{user}/{passwd}:
    get:
      parameters:
      - description: auth or auth-int
        explode: false
        in: path
        name: qop
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: user
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: passwd
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content: {}
          description: Sucessful authentication.
        "401":
          content: {}
          description: Unsuccessful authentication.
      summary: Prompts the user for authorization using Digest Auth.
      tags:
      - Auth
      x-accepts:
      - application/json
      x-tags:
      - tag: Auth
  /digest-auth/{qop}/{user}/{passwd}/{algorithm}:
    get:
      parameters:
      - description: auth or auth-int
        explode: false
        in: path
        name: qop
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: user
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: passwd
        required: true
        schema:
          type: string
        style: simple
      - description: "MD5, SHA-256, SHA-512"
        explode: false
        in: path
        name: algorithm
        required: true
        schema:
          default: MD5
          type: string
        style: simple
      responses:
        "200":
          content: {}
          description: Sucessful authentication.
        "401":
          content: {}
          description: Unsuccessful authentication.
      summary: Prompts the user for authorization using Digest Auth + Algorithm.
      tags:
      - Auth
      x-accepts:
      - application/json
      x-tags:
      - tag: Auth
  /digest-auth/{qop}/{user}/{passwd}/{algorithm}/{stale_after}:
    get:
      description: |
        allow settings the stale_after argument.
      parameters:
      - description: auth or auth-int
        explode: false
        in: path
        name: qop
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: user
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: passwd
        required: true
        schema:
          type: string
        style: simple
      - description: "MD5, SHA-256, SHA-512"
        explode: false
        in: path
        name: algorithm
        required: true
        schema:
          default: MD5
          type: string
        style: simple
      - explode: false
        in: path
        name: stale_after
        required: true
        schema:
          default: never
          type: string
        style: simple
      responses:
        "200":
          content: {}
          description: Sucessful authentication.
        "401":
          content: {}
          description: Unsuccessful authentication.
      summary: Prompts the user for authorization using Digest Auth + Algorithm.
      tags:
      - Auth
      x-accepts:
      - application/json
      x-tags:
      - tag: Auth
...
  /hidden-basic-auth/{user}/{passwd}:
    get:
      parameters:
      - explode: false
        in: path
        name: user
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: passwd
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content: {}
          description: Sucessful authentication.
        "404":
          content: {}
          description: Unsuccessful authentication.
      summary: Prompts the user for authorization using HTTP Basic Auth.
      tags:
      - Auth
      x-accepts:
      - application/json
      x-tags:
      - tag: Auth
...